src: fix probe arguments
Probe arguments weren't working because DTrace USDT accepts a string
with the argument type name, and Systemtap's SDT has only a few types
defined as signed integers. A conversion function was added to map
DTrace types to Systemtap types, but it's always returning uint64 for
now. In the future this function should map those types and return the
appropriate one.